+++ src/tui.rs
@@ -961,7 +961,7 @@ fn source_color(theme: &Theme, source: &Source) -> Color {
 
 #[cfg(test)]
 mod tests {
-    use std::path::PathBuf;
+    use std::{path::PathBuf, sync::OnceLock};
 
     use ratatui::backend::TestBackend;
 
@@ -985,6 +985,7 @@ mod tests {
             action: EntryAction::FocusOrCreateDir,
             source_label: None,
             search_terms: vec![],
+            canonical: OnceLock::new(),
         }
     }